+++ src/appdaemon/appdaemon.yaml
@@ -6,6 +6,5 @@ appdaemon:
   plugins:
     HASS:
       type: hass
-      ha_url: http://192.168.1.100:8123
-      token: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiIzZjQ2MWJkMGQzOTA0ZjlhYWFhZjkwMGY4ZjhiZDQwMSIsImlhdCI6MTYzODM5NDk5OCwiZXhwIjoxOTUzNzU0OTk4fQ.pNTbAOw4FQExiizEz8jkMdHCW3wcpbRnlp0jzlalHz4
-
+      ha_url: https://novabb.duckdns.org:443
+      token: !secret token
