+++ src/appdaemon/apps/apps.yaml
@@ -1,7 +1,3 @@
-hello_world:
-  module: hello
-  class: HelloWorld
-
 model_executor:
   module: model_executor
   class: ModelExecutor
