--- src/appdaemon/apps/hello.py
@@ -1,13 +1,0 @@
-import appdaemon.plugins.hass.hassapi as hass
-
-#
-# Hello World App
-#
-# Args:
-#
-
-
-class HelloWorld(hass.Hass):
-    def initialize(self):
-        self.log("Hello from AppDaemon")
-        self.log("You are now ready to run Apps!")
