--- src/appdaemon/apps/model_creator/4_engine.py
@@ -1,9 +1,0 @@
-#%%
-from hassapi import Hass
-import homeassistant.helpers.event as ev
-import configuration
-
-home = Hass(hassurl="https://novabb.duckdns.org/", token=configuration.auth_token)
-
-states = home.get_states()
-print(states)
